refactor(config): split SpectorProperties into SpectorConfigSource + aggregate POJO
Rename the existing SpectorProperties (config loader/resolver wrapping
CombinedConfiguration) to SpectorConfigSource. Create a new
SpectorProperties as an immutable aggregate root POJO that holds all
typed sub-domain children (MemoryProperties, ProviderProperties,
IngestionProperties, HnswProperties, IvfProperties, SpectrumProperties)
and retains the SpectorConfigSource as an escape hatch for raw key
lookups.

Add SpectorConfigFactory.spectorProperties(SpectorConfigSource) as the
single canonical construction path for the aggregate POJO.

Update all 24 import sites across nucleus, memory, synapse, and bench
modules to use SpectorConfigSource for raw config access.

Refs #757
