+++ nucleus/spector-config/src/main/java/com/spectrayan/spector/config/SpectorConfigSource.java
@@ -1,0 +1,531 @@
+/*
+ * Copyright 2026 Spectrayan
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.spectrayan.spector.config;
+
+import com.spectrayan.spector.commons.error.SpectorValidationException;
+import com.spectrayan.spector.commons.error.ErrorCode;
+import com.spectrayan.spector.config.error.SpectorConfigNotFoundException;
+
+import org.apache.commons.configuration2.CombinedConfiguration;
+import org.apache.commons.configuration2.Configuration;
+import org.apache.commons.configuration2.MapConfiguration;
+import org.apache.commons.configuration2.PropertiesConfiguration;
+import org.apache.commons.configuration2.YAMLConfiguration;
+import org.apache.commons.configuration2.ex.ConfigurationException;
+import org.apache.commons.configuration2.interpol.DefaultLookups;
+import org.apache.commons.configuration2.tree.OverrideCombiner;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import java.io.IOException;
+import java.io.InputStream;
+import java.io.Reader;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.time.Duration;
+import java.util.Iterator;
+import java.util.Map;
+import java.util.Objects;
+import java.util.Properties;
+import com.spectrayan.spector.commons.error.ErrorCode;
+
+/**
+ * Spring Boot-style hierarchical configuration loader for Spector.
+ *
+ * <p>Provides a single point of entry for all configuration across modules.
+ * Uses Apache Commons Configuration 2 under the hood with a
+ * {@link CombinedConfiguration} and {@link OverrideCombiner} to layer
+ * multiple configuration sources.</p>
+ *
+ * <h3>Resolution Order (highest priority wins)</h3>
+ * <ol>
+ *   <li>Programmatic overrides (via {@link Builder#override(String, Object)})</li>
+ *   <li>System properties ({@code -Dspector.memory.dimensions=768})</li>
+ *   <li>System environment variables ({@code SPECTOR_MEMORY_DIMENSIONS=768})</li>
+ * </ul>
+ *
+ * <p>Examples:</p>
+ * <pre>{@code
+ * SpectorConfigSource props = SpectorConfigSource.load();
+ * int dims = props.getInt("spector.memory.dimensions", 768);
+ * Path path = props.getPath("spector.memory.persistence-path", Path.of(".spector/memory"));
+ * }</pre>
+ *
+ * <h3>Environment Variable Mapping</h3>
+ * <ul>
+ *   <li>{@code spector.memory.dimensions} → {@code SPECTOR_MEMORY_DIMENSIONS}</li>
+ * </ol>
+ *
+ * <h3>Usage</h3>
+ * <pre>{@code
+ *   // Auto-detect (loads from working dir + classpath defaults)
+ *   SpectorConfigSource props = SpectorConfigSource.load();
+ *
+ *   // With profile
+ *   SpectorConfigSource props = SpectorConfigSource.load("production");
+ *
+ *   // From explicit file
+ *   SpectorConfigSource props = SpectorConfigSource.load(Path.of("/etc/spector/spector.yml"));
+ *
+ *   // Typed access
+ *   int dims = props.getInt("spector.memory.dimensions", 768);
+ *   String model = props.getString("spector.provider.embedding.model", "nomic-embed-text");
+ *   Duration timeout = props.getDuration("spector.provider.embedding.timeout", Duration.ofSeconds(30));
+ * }</pre>
+ *
+ * <h3>Environment Variable Mapping</h3>
+ * <p>Dot-notation keys are mapped to environment variables by uppercasing and
+ * replacing dots/hyphens with underscores:</p>
+ * <ul>
+ *   <li>{@code spector.memory.dimensions} → {@code SPECTOR_MEMORY_DIMENSIONS}</li>
+ *   <li>{@code spector.hnsw.ef-construction} → {@code SPECTOR_HNSW_EF_CONSTRUCTION}</li>
+ * </ul>
+ */
+public final class SpectorConfigSource {
+
+    private static final Logger log = LoggerFactory.getLogger(SpectorConfigSource.class);
+
+    /** Default config file name in working directory. */
+    private static final String DEFAULT_CONFIG_FILE = "spector.yml";
+
+    /** Default config file name on classpath (bundled in JAR). */
+    private static final String CLASSPATH_DEFAULTS = "spector-defaults.yml";
+
+    /** Profile config file pattern. */
+    private static final String PROFILE_PATTERN = "spector-%s.yml";
+
+    private final CombinedConfiguration config;
+
+    private SpectorConfigSource(CombinedConfiguration config) {
+        if (config == null) { throw new SpectorValidationException(ErrorCode.ARGUMENT_NULL, "config"); } this.config = config;
+    }
+
+    // ─────────────── Static Factory Methods ───────────────
+
+    /**
+     * Loads configuration with auto-detection.
+     * <p>Checks for a {@code spector.profile} system property or
+     * {@code SPECTOR_PROFILE} environment variable to determine the active profile.</p>
+     */
+    public static SpectorConfigSource load() {
+        String profile = System.getProperty("spector.profile",
+                System.getenv().getOrDefault("SPECTOR_PROFILE", null));
+        return new Builder().profile(profile).build();
+    }
+
+    /**
+     * Loads configuration from classpath defaults only — no filesystem discovery.
+     * <p>Useful for tests that should not be affected by a {@code spector.yml}
+     * file in the working directory.</p>
+     */
+    public static SpectorConfigSource loadClasspathOnly() {
+        return new Builder().skipFilesystemDiscovery(true).build();
+    }
+
+    /**
+     * Loads configuration with the specified profile.
+     *
+     * @param profile the active profile name (e.g., "dev", "production"), or null for none
+     */
+    public static SpectorConfigSource load(String profile) {
+        return new Builder().profile(profile).build();
+    }
+
+    /**
+     * Loads configuration from an explicit file path.
+     *
+     * @param configFile path to the primary configuration file
+     */
+    public static SpectorConfigSource load(Path configFile) {
+        return new Builder().configFile(configFile).build();
+    }
+
+    /**
+     * Creates a new builder for fine-grained control over configuration loading.
+     */
+    public static Builder builder() {
+        return new Builder();
+    }
+
+    // ─────────────── Typed Accessors ───────────────
+
+    /**
+     * Returns the string value for the given key, or {@code null} if not found.
+     */
+    public String getString(String key) {
+        return resolveWithEnv(key, null);
+    }
+
+    /**
+     * Returns the string value for the given key, or the default if not found.
+     */
+    public String getString(String key, String defaultValue) {
+        String value = resolveWithEnv(key, null);
+        return value != null ? value : defaultValue;
+    }
+
+    /**
+     * Returns the integer value for the given key, or the default if not found.
+     */
+    public int getInt(String key, int defaultValue) {
+        String value = resolveWithEnv(key, null);
+        if (value != null) {
+            try {
+                return Integer.parseInt(value);
+            } catch (NumberFormatException e) {
+                log.warn("Invalid integer for key '{}': '{}', using default {}", key, value, defaultValue);
+            }
+        }
+        return config.getInt(key, defaultValue);
+    }
+
+    /**
+     * Returns the long value for the given key, or the default if not found.
+     */
+    public long getLong(String key, long defaultValue) {
+        String value = resolveWithEnv(key, null);
+        if (value != null) {
+            try {
+                return Long.parseLong(value);
+            } catch (NumberFormatException e) {
+                log.warn("Invalid long for key '{}': '{}', using default {}", key, value, defaultValue);
+            }
+        }
+        return config.getLong(key, defaultValue);
+    }
+
+    /**
+     * Returns the boolean value for the given key, or the default if not found.
+     */
+    public boolean getBoolean(String key, boolean defaultValue) {
+        String value = resolveWithEnv(key, null);
+        if (value != null) {
+            return Boolean.parseBoolean(value);
+        }
+        return config.getBoolean(key, defaultValue);
+    }
+
+    /**
+     * Returns the double value for the given key, or the default if not found.
+     */
+    public double getDouble(String key, double defaultValue) {
+        String value = resolveWithEnv(key, null);
+        if (value != null) {
+            try {
+                return Double.parseDouble(value);
+            } catch (NumberFormatException e) {
+                log.warn("Invalid double for key '{}': '{}', using default {}", key, value, defaultValue);
+            }
+        }
+        return config.getDouble(key, defaultValue);
+    }
+
+    /**
+     * Returns the float value for the given key, or the default if not found.
+     */
+    public float getFloat(String key, float defaultValue) {
+        String value = resolveWithEnv(key, null);
+        if (value != null) {
+            try {
+                return Float.parseFloat(value);
+            } catch (NumberFormatException e) {
+                log.warn("Invalid float for key '{}': '{}', using default {}", key, value, defaultValue);
+            }
+        }
+        return config.getFloat(key, defaultValue);
+    }
+
+    /**
+     * Returns a {@link Duration} parsed from the given key.
+     * <p>Supports formats: {@code 30s}, {@code 5m}, {@code 1h}, {@code 500ms},
+     * or ISO-8601 ({@code PT30S}).</p>
+     */
+    public Duration getDuration(String key, Duration defaultValue) {
+        String value = getString(key);
+        if (value == null || value.isBlank()) return defaultValue;
+        try {
+            return parseDuration(value, key, defaultValue);
+        } catch (Exception customErr) {
+            try {
+                return config.getDuration(key, defaultValue);
+            } catch (Exception apacheErr) {
+                log.warn("Invalid duration for key '{}': '{}', using default {}", key, value, defaultValue);
+                return defaultValue;
+            }
+        }
+    }
+
+    /**
+     * Returns a {@link Path} for the given key, or the default if not found.
+     */
+    public Path getPath(String key, Path defaultValue) {
+        String value = getString(key);
+        if (value == null || value.isBlank()) return defaultValue;
+        return Path.of(value);
+    }
+
+    /**
+     * Returns an enum value for the given key, or the default if not found.
+     */
+    public <E extends Enum<E>> E getEnum(String key, Class<E> enumType, E defaultValue) {
+        String value = getString(key);
+        if (value == null || value.isBlank()) return defaultValue;
+        try {
+            return Enum.valueOf(enumType, value.toUpperCase().replace('-', '_'));
+        } catch (IllegalArgumentException e) {
+            log.warn("Invalid {} value for key '{}': '{}', using default {}",
+                    enumType.getSimpleName(), key, value, defaultValue);
+            return defaultValue;
+        }
+    }
+
+    /**
+     * Returns a view of this configuration scoped to the given prefix.
+     * <p>Example: {@code subset("spector.embedding")} returns a view where
+     * the key {@code "model"} maps to {@code "spector.embedding.model"}.</p>
+     */
+    public SpectorConfigSource subset(String prefix) {
+        CombinedConfiguration sub = new CombinedConfiguration(new OverrideCombiner());
+        sub.addConfiguration(config.subset(prefix));
+        return new SpectorConfigSource(sub);
+    }
+
+    /**
+     * Checks if the configuration contains the given key.
+     */
+    public boolean containsKey(String key) {
+        return resolveWithEnv(key, null) != null || config.containsKey(key);
+    }
+
+    /**
+     * Returns all keys in the configuration.
+     */
+    public Iterator<String> getKeys() {
+        return config.getKeys();
+    }
+
+    // ─────────────── Environment Variable Resolution ───────────────
+
+    /**
+     * Resolves a key by first checking system properties, then environment
+     * variables (with dot-to-underscore mapping), then the configuration.
+     */
+    private String resolveWithEnv(String key, String defaultValue) {
+        // 1. System property
+        String sysProp = System.getProperty(key);
+        if (sysProp != null) return sysProp;
+
+        // 2. Environment variable (spector.memory.dimensions → SPECTOR_MEMORY_DIMENSIONS)
+        String envKey = key.toUpperCase().replace('.', '_').replace('-', '_');
+        String envValue = System.getenv(envKey);
+        if (envValue != null) return envValue;
+
+        // 3. Configuration file (variable interpolation handled natively by Commons Configuration 2)
+        String configValue = config.getString(key, null);
+        return configValue != null ? configValue : defaultValue;
+    }
+
+    // ─────────────── Duration Parsing ───────────────
+
+    private static Duration parseDuration(String value, String key, Duration defaultValue) {
+        // Try ISO-8601 first (PT30S, PT5M, etc.)
+        if (value.startsWith("PT") || value.startsWith("pt")) {
+            return Duration.parse(value);
+        }
+
+        // Human-readable: 30s, 5m, 1h, 500ms
+        String trimmed = value.trim().toLowerCase();
+        if (trimmed.endsWith("ms")) {
+            return Duration.ofMillis(Long.parseLong(trimmed.substring(0, trimmed.length() - 2).trim()));
+        } else if (trimmed.endsWith("s")) {
+            return Duration.ofSeconds(Long.parseLong(trimmed.substring(0, trimmed.length() - 1).trim()));
+        } else if (trimmed.endsWith("m")) {
+            return Duration.ofMinutes(Long.parseLong(trimmed.substring(0, trimmed.length() - 1).trim()));
+        } else if (trimmed.endsWith("h")) {
+            return Duration.ofHours(Long.parseLong(trimmed.substring(0, trimmed.length() - 1).trim()));
+        }
+
+        // Try as seconds if just a number
+        return Duration.ofSeconds(Long.parseLong(trimmed));
+    }
+
+    // ─────────────── Builder ───────────────
+
+    /**
+     * Builder for fine-grained control over {@link SpectorConfigSource} construction.
+     */
+    public static class Builder {
+        private String profile;
+        private Path configFile;
+        private boolean skipFilesystem;
+        private final Properties overrides = new Properties();
+
+        private Builder() {}
+
+        /** Sets the active profile (e.g., "dev", "production"). */
+        public Builder profile(String profile) {
+            this.profile = profile;
+            return this;
+        }
+
+        /** Sets an explicit configuration file path. */
+        public Builder configFile(Path configFile) {
+            this.configFile = configFile;
+            return this;
+        }
+
+        /** If true, skip loading from working-directory files (spector.yml, spector.properties). */
+        Builder skipFilesystemDiscovery(boolean skip) {
+            this.skipFilesystem = skip;
+            return this;
+        }
+
+        /** Adds a programmatic override. */
+        public Builder override(String key, Object value) {
+            overrides.setProperty(key, String.valueOf(value));
+            return this;
+        }
+
+        /** Adds multiple programmatic overrides. */
+        public Builder overrides(Map<String, ?> overrides) {
+            overrides.forEach((k, v) -> this.overrides.setProperty(k, String.valueOf(v)));
+            return this;
+        }
+
+        /**
+         * Builds the {@link SpectorConfigSource} instance.
+         *
+         * <p>Layer order (first added = highest priority with OverrideCombiner):</p>
+         * <ol>
+         *   <li>Programmatic overrides</li>
+         *   <li>Profile-specific YAML (if profile set)</li>
+         *   <li>User config file (explicit path or spector.yml in working dir)</li>
+         *   <li>Classpath defaults (spector-defaults.yml)</li>
+         * </ol>
+         */
+        public SpectorConfigSource build() {
+            CombinedConfiguration combined = new CombinedConfiguration(new OverrideCombiner());
+            configureInterpolator(combined);
+
+            // 1. Programmatic overrides (highest priority)
+            if (!overrides.isEmpty()) {
+                @SuppressWarnings({"unchecked", "rawtypes"})
+                MapConfiguration overrideConfig = new MapConfiguration((Map) overrides);
+                combined.addConfiguration(overrideConfig, "overrides");
+                log.debug("[SpectorConfigSource] Added {} programmatic overrides", overrides.size());
+            }
+
+            // 2. Profile-specific file
+            if (profile != null && !profile.isBlank()) {
+                String profileFileName = String.format(PROFILE_PATTERN, profile);
+                loadFileIfExists(combined, Path.of(profileFileName), "profile-" + profile);
+                loadClasspathYaml(combined, profileFileName, "classpath-profile-" + profile);
+            }
+
+            // 3. User config file
+            if (configFile != null) {
+                loadFileOrFail(combined, configFile, "user-config");
+            } else if (!skipFilesystem) {
+                // Try spector.yml in working directory
+                loadFileIfExists(combined, Path.of(DEFAULT_CONFIG_FILE), "user-config");
+                // Also try spector.properties as fallback
+                loadPropertiesIfExists(combined, Path.of("spector.properties"), "user-properties");
+            }
+
+            // 4. Classpath defaults (lowest priority)
+            loadClasspathYaml(combined, CLASSPATH_DEFAULTS, "classpath-defaults");
+
+            log.info("[SpectorConfigSource] Loaded {} configuration sources{}",
+                    combined.getNumberOfConfigurations(),
+                    profile != null ? " (profile: " + profile + ")" : "");
+
+            return new SpectorConfigSource(combined);
+        }
+
+        // ─── File Loading Helpers ───
+
+        private void loadFileIfExists(CombinedConfiguration combined, Path path, String name) {
+            if (Files.isRegularFile(path)) {
+                try {
+                    String fileName = path.getFileName().toString();
+                    if (fileName.endsWith(".yml") || fileName.endsWith(".yaml")) {
+                        YAMLConfiguration yaml = new YAMLConfiguration();
+                        configureInterpolator(yaml);
+                        try (Reader reader = Files.newBufferedReader(path)) {
+                            yaml.read(reader);
+                        }
+                        combined.addConfiguration(yaml, name);
+                        log.debug("[SpectorConfigSource] Loaded YAML: {}", path.toAbsolutePath());
+                    } else if (fileName.endsWith(".properties")) {
+                        PropertiesConfiguration props = new PropertiesConfiguration();
+                        configureInterpolator(props);
+                        try (Reader reader = Files.newBufferedReader(path)) {
+                            props.read(reader);
+                        }
+                        combined.addConfiguration(props, name);
+                        log.debug("[SpectorConfigSource] Loaded properties: {}", path.toAbsolutePath());
+                    }
+                } catch (ConfigurationException | IOException e) {
+                    log.warn("[SpectorConfigSource] Failed to load {}: {}", path, e.getMessage());
+                }
+            }
+        }
+
+        private void loadFileOrFail(CombinedConfiguration combined, Path path, String name) {
+            if (!Files.isRegularFile(path)) {
+                throw new SpectorConfigNotFoundException(path.toAbsolutePath().toString());
+            }
+            loadFileIfExists(combined, path, name);
+        }
+
+        private void loadPropertiesIfExists(CombinedConfiguration combined, Path path, String name) {
+            if (Files.isRegularFile(path)) {
+                try {
+                    PropertiesConfiguration props = new PropertiesConfiguration();
+                    configureInterpolator(props);
+                    try (Reader reader = Files.newBufferedReader(path)) {
+                        props.read(reader);
+                    }
+                    combined.addConfiguration(props, name);
+                    log.debug("[SpectorConfigSource] Loaded properties: {}", path.toAbsolutePath());
+                } catch (ConfigurationException | IOException e) {
+                    log.warn("[SpectorConfigSource] Failed to load {}: {}", path, e.getMessage());
+                }
+            }
+        }
+
+        private void loadClasspathYaml(CombinedConfiguration combined, String resource, String name) {
+            try (InputStream is = SpectorConfigSource.class.getClassLoader().getResourceAsStream(resource)) {
+                if (is != null) {
+                    YAMLConfiguration yaml = new YAMLConfiguration();
+                    configureInterpolator(yaml);
+                    yaml.read(new java.io.InputStreamReader(is));
+                    combined.addConfiguration(yaml, name);
+                    log.debug("[SpectorConfigSource] Loaded classpath: {}", resource);
+                }
+            } catch (ConfigurationException | IOException e) {
+                log.warn("[SpectorConfigSource] Failed to load classpath {}: {}", resource, e.getMessage());
+            }
+        }
+
+        private static void configureInterpolator(Configuration config) {
+            if (config instanceof org.apache.commons.configuration2.AbstractConfiguration ac) {
+                ac.getInterpolator().addDefaultLookup(DefaultLookups.ENVIRONMENT.getLookup());
+                ac.getInterpolator().addDefaultLookup(DefaultLookups.SYSTEM_PROPERTIES.getLookup());
+            }
+        }
+    }
+}
