+++ nucleus/spector-config/src/test/java/com/spectrayan/spector/config/SpectorConfigFactoryTest.java
@@ -32,7 +32,7 @@ class SpectorConfigFactoryTest {
 
     @Test
     void hnswProperties_fromClasspath() {
-        var hnsw = SpectorConfigFactory.hnswProperties(SpectorProperties.load());
+        var hnsw = SpectorConfigFactory.hnswProperties(SpectorConfigSource.load());
 
         assertThat(hnsw.m()).isEqualTo(16);
         assertThat(hnsw.efConstruction()).isEqualTo(200);
@@ -41,7 +41,7 @@ class SpectorConfigFactoryTest {
 
     @Test
     void ivfProperties_fromClasspath() {
-        var ivf = SpectorConfigFactory.ivfProperties(SpectorProperties.load());
+        var ivf = SpectorConfigFactory.ivfProperties(SpectorConfigSource.load());
 
         assertThat(ivf.nlist()).isEqualTo(0);
         assertThat(ivf.nprobe()).isEqualTo(0);
@@ -50,7 +50,7 @@ class SpectorConfigFactoryTest {
 
     @Test
     void spectrumProperties_fromClasspath() {
-        var spectrum = SpectorConfigFactory.spectrumProperties(SpectorProperties.load());
+        var spectrum = SpectorConfigFactory.spectrumProperties(SpectorConfigSource.load());
 
         assertThat(spectrum.nCentroids()).isEqualTo(256);
         assertThat(spectrum.nProbe()).isEqualTo(16);
@@ -61,7 +61,7 @@ class SpectorConfigFactoryTest {
 
     @Test
     void embeddingProperties_fromClasspath() {
-        var embed = SpectorConfigFactory.embeddingProperties(SpectorProperties.load());
+        var embed = SpectorConfigFactory.embeddingProperties(SpectorConfigSource.load());
 
         assertThat(embed.model()).isEqualTo("nomic-embed-text");
         assertThat(embed.baseUrl()).isEqualTo("http://localhost:11434");
@@ -72,7 +72,7 @@ class SpectorConfigFactoryTest {
 
     @Test
     void memoryProperties_fromClasspath() {
-        var memory = SpectorConfigFactory.memoryProperties(SpectorProperties.load());
+        var memory = SpectorConfigFactory.memoryProperties(SpectorConfigSource.load());
 
         assertThat(memory.enabled()).isFalse();
         assertThat(memory.persistenceMode()).isEqualTo(PersistenceMode.DISK);
@@ -86,7 +86,7 @@ class SpectorConfigFactoryTest {
 
     @Test
     void ingestionProperties_fromClasspath() {
-        var ingestion = SpectorConfigFactory.ingestionProperties(SpectorProperties.load());
+        var ingestion = SpectorConfigFactory.ingestionProperties(SpectorConfigSource.load());
 
         assertThat(ingestion.rootDirectory()).isEqualTo(Path.of("."));
         assertThat(ingestion.filePattern()).isEqualTo("**/*.md");
@@ -97,7 +97,7 @@ class SpectorConfigFactoryTest {
 
     @Test
     void memoryProperties_flexibleCaseInsensitiveEnums() {
-        SpectorProperties props = SpectorProperties.builder()
+        SpectorConfigSource props = SpectorConfigSource.builder()
                 .override("spector.memory.persistence-mode", "in-memory")
                 .override("spector.memory.default-ingestion-tier", "semantic")
                 .override("spector.memory.hnsw-prefilter", "Enabled")
